Ajuste CrudService e testes unitários de PersonagemService

+++ POTTER-service/src/main/java/com/challenge/CrudService.java
@@ -1,6 +1,7 @@
 package com.challenge;
 
 import com.challenge.entity.BaseEntity;
+import com.challenge.exception.ParametroInvalidoException;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -24,18 +25,26 @@ public abstract class CrudService<T extends BaseEntity> {
     }
 
     public T findById(Long id) {
+        if(id == null)
+            throw new ParametroInvalidoException("Id fornecido não pode ser nulo");
         return this.jpaRepository.findById(id).orElse(null);
     }
 
     public T save(T baseEntity) {
+        if(baseEntity == null)
+            throw new ParametroInvalidoException("Entidade fornecida não pode ser nula");
         return this.jpaRepository.save(baseEntity);
     }
 
     public void delete(T baseEntity) {
+        if(baseEntity == null)
+            throw new ParametroInvalidoException("Entidade fornecida não pode ser nula");
         this.jpaRepository.delete(baseEntity);
     }
 
     public void deleteById(Long id) {
+        if(id == null)
+            throw new ParametroInvalidoException("Id fornecido não pode ser nulo");
         this.jpaRepository.deleteById(id);
     }
